+++ tests/test_components/test_mcms.py
@@ -1,15 +1,13 @@
 import pytest
+from requests.exceptions import InvalidURL
 from flexiznam import mcms
+
 
 USERNAME = "ab8"
 
 
-@pytest.mark.slow
-def test_download_mouse():
-    ret = mcms.main.download_mouse_info(username=USERNAME, mouse_name="PZAJ2.1c")
-    assert ret
-
-
-@pytest.mark.slow
 def test_get_mouse_df():
-    mcms.get_mouse_df(mouse_name="PZAJ2.1c", username=USERNAME)
+    md = mcms.get_mouse_df(mouse_name="PZAJ2.1c", username=USERNAME)
+    assert md["mcms_id"] == 1431106
+    with pytest.raises(InvalidURL):
+        mcms.get_mouse_df(mouse_name="wrongname", username=USERNAME)