+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add `enforce_dataset_types` option to `config.yml`. This let the user decide if the
   dataset type must be defined in the config file or can be freely changed.
 - `Dataset.from_flexilims` accepts `id` or `name`.
+- Use `pymcms` to get mouse data from MCMS
 
 ### Bugfixes
 - `add_genealogy` now works with scanimage datasets